move helper functions to utils, fix tests
